Penup/Pendown
You can now press the "e" key, to start and stop drawing. sometimes it doesn't really stop or start because the loop already ran through again and set the value back to what it was before because the key was still pressed. I'm gonna look around if I find a way to fix that.